requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.1
billiard==3.6.1.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
croniter==0.3.31
Django==2.2.7
django-authtools==1.7.0
django-oauth-toolkit==1.2.0
django-rq==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==6.2.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
rq==1.2.2
rq-scheduler==0.9.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
vk-api==11.7.0
zipp==0.6.0
